added --noupdate option to use previously retrieved (ossibly locally modified) sources Original: 52679a100cdb9ad86cf3f7d06a9dfd0c89b65176

+++ lalapps/src/pulsar/hough/src2/EinsteinAtHome/eah_build2.sh
@@ -78,6 +78,8 @@ for i; do
 	    rebuild_boinc=""
 	    rebuild_lal=""
 	    rebuild="" ;;
+        --noupdate)
+            noupdate=true ;;
         --nohough)
 	    nohough=true ;;
 	--gc-opt)
@@ -179,6 +181,7 @@ for i; do
 	    echo "  --appname=<name>  set an application name (only used in --release builds, defaults to einstein_S5GC1HF)"
 	    echo "  --appversion=N.NN set an application version (only used in --release builds, defaults to 0.00)"
 	    echo "  --norebuild       disables --rebuild on --release. DANGEROUS! Use only for testing the build script"
+	    echo "  --noupdate        use previously retrieved (ossibly locally modified) sources, doesn't need internet"
 	    echo "  --nohough         don't build HierarchicalSearch from pulsar/hough/src2, just build HierarchSearchGCT"
 	    echo "  --help            show this message and exit"
 	    exit ;;
@@ -326,7 +329,7 @@ log_and_do cd "$SOURCE"
 
 if test -z "$rebuild" && pkg-config --exists gsl; then
     log_and_show "using existing gsl"
-else
+elif test -z "$noupdate"; then
     log_and_show "retrieving $gsl"
     download http://www.aei.mpg.de/~repr/EaH_packages $gsl.tar.gz
     log_and_do tar xzf "$gsl.tar.gz"
@@ -334,13 +337,13 @@ fi
 
 if test -z "$rebuild" && pkg-config --exists fftw3 fftw3f; then
     log_and_show "using existing fftw"
-else
+elif test -z "$noupdate"; then
     log_and_show "retrieving $fftw"
     download ftp://ftp.fftw.org/pub/fftw $fftw.tar.gz
     log_and_do tar xzf "$fftw.tar.gz"
 fi
 
-if test -n "$build_binutils" -a -n "$rebuild_binutils"; then
+if test -n "$build_binutils" -a -n "$rebuild_binutils" -a -z "$noupdate"; then
     log_and_show "retrieving $binutils"
     download http://www.aei.mpg.de/~bema $binutils.tar.gz
     log_and_do rm -rf "$binutils"
@@ -349,7 +352,7 @@ if test -n "$build_binutils" -a -n "$rebuild_binutils"; then
 #    log_and_do mv $binutils/bfd/Makefile.tmp $binutils/bfd/Makefile.am
 fi
 
-if test -z "$rebuild_boinc" -a -d "$SOURCE/boinc" ; then
+if test -n "$noupdate" -o -z "$rebuild_boinc" -a -d "$SOURCE/boinc"; then
     log_and_show "using existing boinc source"
 else
     log_and_show "retrieving boinc"
